personal/graphs/lightgcn/requirements_cuda.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.13.1+cu116 -f https://download.pytorch.org/whl/torch_stable.html
torch-scatter==2.1.0 -f https://data.pyg.org/whl/torch-1.13.0%2Bcu116.html
torch-sparse==0.6.16 -f https://data.pyg.org/whl/torch-1.13.0%2Bcu116.html
torch-cluster==1.6.0 -f https://data.pyg.org/whl/torch-1.13.0%2Bcu116.html
torch-spline-conv==1.2.1 -f https://data.pyg.org/whl/torch-1.13.0%2Bcu116.html
torch-geometric==2.2.0
torchvision==0.14.1
torchaudio==0.13.1